requirements.txt
python-chess>=1.9.0
PyQt6>=6.4.0
matplotlib>=3.7.0
pytest>=7.0.0
requests>=2.31.0
